+++ BK7231Flasher/Flashers/XR809Flasher.cs
@@ -1,37 +1,16 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Drawing;
 using System.IO;
-using System.IO.Ports;
-using System.Text;
 using System.Threading;
 
 namespace BK7231Flasher
 {
-    public class XR809Flasher : BaseFlasher
+    public class XR809Flasher : XRBaseFlasher
     {
-        // =====================================================================
-        // Constants
-        // =====================================================================
-
-        const int  XR_ROM_BAUD         = 115200;    // BROM default after reset
-        const int  XR_SECTOR_SIZE      = 0x200;     // 512 bytes
         const int  XR_ERASE_BLOCK_SIZE = 0x10000;   // 64 KB erase block used by XR809 block erase
-        const int  XR_WRITE_CHUNK_SIZE = 0x4000;    // 16 KB, 32 sectors per write command
-        const int  XR_READ_RETRY_COUNT  = 3;
-        const int  XR_WRITE_RETRY_COUNT = 2;
-        const int  XR_SYNC_RETRY_COUNT  = 10;
-        const int  XR_UPGRADE_SETTLE_MS = 120;
-        const int  XR_MAX_BROM_PAYLOAD  = XR_WRITE_CHUNK_SIZE;
         const int  XR_STUB_LOAD_ADDR    = 0x00010000;
         const int  XR_STUB_ENTRY_ADDR   = 0x00010101;
-
-        // BROM packet byte constants
-        const byte BROM_HOST_FLAGS     = 0x04;      // host→device flags byte
-        const byte BROM_HOST_QUERY     = 0x00;      // header byte 5 for simple query packets
-        const byte BROM_HOST_PAYLOAD   = 0x01;      // header byte 5 for payload-bearing packets
-        const byte BROM_FLAG_ERROR     = 0x01;      // bit 0 of device→host flags byte = command failed
 
         // Opcodes
         const byte OP_WRITE_MEMORY     = 0x09;
@@ -63,129 +42,20 @@ namespace BK7231Flasher
             { 0xA5F85A07, "wlan_sdd" },
         };
 
-        // =====================================================================
-        // Fields
-        // =====================================================================
-        MemoryStream readResult;
-        byte[]       flashID;
-        int          flashSizeBytes = 2 * 0x100000;
-        byte         bromVersion    = 0x00;
         bool         ramStubLoaded;
         bool         useExtendedFlashOpcodes;
-
-        // =====================================================================
-        // Response / header structures
-        // =====================================================================
-        struct BROMResponse
-        {
-            public byte   Flags;
-            public byte   BromVersion;
-            public ushort Checksum;         // present in header but never validated by the BROM
-            public int    PayloadLength;
-            public byte[] Payload;
-            public bool   IsError => (Flags & BROM_FLAG_ERROR) != 0;
-        }
-
-        struct XRSectionHeader
-        {
-            public uint   Magic;
-            public uint   Version;
-            public ushort HeaderChecksum;
-            public ushort DataChecksum;
-            public uint   DataSize;
-            public uint   LoadAddr;
-            public uint   Entry;
-            public uint   BodyLen;
-            public uint   Attribute;
-            public uint   NextAddr;
-            public uint   SectionId;
-        }
-
-        // =====================================================================
-        // Constructor
-        // =====================================================================
         public XR809Flasher(CancellationToken ct) : base(ct)
         {
         }
 
-
-        int GetReadWaitMs()
+        protected override Dictionary<uint, string> SectionNames
         {
-            int baud = serial?.BaudRate ?? XR_ROM_BAUD;
-            switch (baud)
-            {
-                case 9600:
-                    return int.MaxValue;
-                case 115200:
-                case 230400:
-                case 460800:
-                    return 500;
-                case 921600:
-                case 1000000:
-                    return 100;
-                case 1500000:
-                case 2000000:
-                    return 70;
-                case 3000000:
-                    return 30;
-                default:
-                    return 500;
-            }
+            get { return SECTION_NAMES; }
         }
 
-        int GetReadTimeoutMs()
+        protected override string FormatHeaderIncompleteMessage(byte[] header)
         {
-            int waitMs = GetReadWaitMs();
-            if (waitMs == int.MaxValue)
-                return int.MaxValue;
-            int total = waitMs * 40;
-            return total < 1000 ? 1000 : total;
-        }
-
-        // Easy Flasher passes XR809 startSector/sectors in the shared public
-        // 4 KB framework unit used by several platform front-ends. Convert that
-        // contract to byte addresses here, then later to the native 512-byte
-        // XR transport sectors used on the wire.
-        static int PublicSectorIndexToByteAddress(int startSector)
-        {
-            return startSector * BK7231Flasher.SECTOR_SIZE;
-        }
-
-        static int PublicSectorCountToByteLength(int sectors)
-        {
-            return sectors * BK7231Flasher.SECTOR_SIZE;
-        }
-
-        bool IsCustomRawWriteRequest(int startSector, int sectors, WriteMode rwMode)
-        {
-            if (rwMode != WriteMode.OnlyWrite)
-                return false;
-
-            int fullFlashFrameworkSectors = BK7231Flasher.FLASH_SIZE / BK7231Flasher.SECTOR_SIZE;
-            return startSector != 0 || sectors != fullFlashFrameworkSectors;
-        }
-
-        static bool StartsWithAwihMagic(byte[] data)
-        {
-            return data != null && data.Length >= 4 && BitConverter.ToUInt32(data, 0) == 0x48495741;
-        }
-
-        // =====================================================================
-        // Checksum
-        //
-        // Returns the 16-bit additive checksum used by XR packet preimages
-        // and XR .img section validation. The running sum wraps naturally at
-        // 16 bits and the final value is bitwise inverted.
-        // =====================================================================
-        static ushort ComputeChecksum(byte[] data, int offset, int count)
-        {
-            ushort sum = 0;
-            int i = 0;
-            for (; i + 1 < count; i += 2)
-                sum += (ushort)(data[offset + i] | (data[offset + i + 1] << 8));
-            if ((count & 1) != 0)
-                sum += data[offset + count - 1];
-            return (ushort)(~sum);
+            return $"BROM header incomplete ({header?.Length ?? 0}/12 bytes): {FormatHexSnippet(header)}.";
         }
 
         // =====================================================================
@@ -196,7 +66,7 @@ namespace BK7231Flasher
         // requestType is header byte 5: 0 for simple query packets and 1 for
         // payload-bearing commands.
         // =====================================================================
-        byte[] BuildBromPacket(byte opcode, byte[] payloadPre, byte[] payloadWire, byte requestType = BROM_HOST_PAYLOAD)
+        new byte[] BuildBromPacket(byte opcode, byte[] payloadPre, byte[] payloadWire, byte requestType = BROM_HOST_PAYLOAD)
         {
             if (payloadPre  == null) payloadPre  = new byte[0];
             if (payloadWire == null) payloadWire = new byte[0];
@@ -371,7 +241,7 @@ namespace BK7231Flasher
         }
 
         // ChangeBaud packet: 4-byte payload = (baud | 0x03000000) BE.
-        byte[] BuildChangeBaudPacket(int newBaud)
+        protected override byte[] BuildChangeBaudPacket(int newBaud)
         {
             int val = newBaud | unchecked((int)0x03000000);
             var pre  = new byte[4];
@@ -385,295 +255,6 @@ namespace BK7231Flasher
             wire[2] = (byte)((val >>  8) & 0xFF);
             wire[3] = (byte)( val        & 0xFF);
             return BuildBromPacket(OP_CHANGE_BAUD, pre, wire, BROM_HOST_PAYLOAD);
-        }
-
-        // =====================================================================
-        // Port setup / teardown
-        // =====================================================================
-        bool DoGenericSetup()
-        {
-            addLog("Now is: " + DateTime.Now.ToLongDateString()
-                              + " " + DateTime.Now.ToLongTimeString()
-                              + "." + Environment.NewLine);
-            addLog("Flasher mode: " + chipType + Environment.NewLine);
-            addLog("Going to open port: " + serialName + "." + Environment.NewLine);
-            try
-            {
-                serial = new SerialPort(serialName, XR_ROM_BAUD);
-                serial.ReadTimeout     = 250;
-                serial.WriteTimeout    = 5000;
-                serial.ReadBufferSize  = 1024 * 1024;
-                serial.WriteBufferSize = 1024 * 1024;
-                serial.Open();
-                serial.DiscardInBuffer();
-                serial.DiscardOutBuffer();
-            }
-            catch (Exception ex)
-            {
-                addErrorLine("Port setup failed: " + ex.Message);
-                return false;
-            }
-            addLogLine("Port ready!");
-            return true;
-        }
-
-        // Sends ChangeBaud(115200) to the device before closing so it returns to the
-        // default baud and doesn't need a power cycle to reconnect next session.
-        // Best-effort — any failure is swallowed; base.closePort() is always called.
-        public override void closePort()
-        {
-            if (serial != null && serial.IsOpen && serial.BaudRate != XR_ROM_BAUD)
-            {
-                try
-                {
-                    byte[] pkt = BuildChangeBaudPacket(XR_ROM_BAUD);
-                    serial.Write(pkt, 0, pkt.Length);
-                    serial.BaseStream.Flush();
-                    Thread.Sleep(60);
-                    serial.BaudRate = XR_ROM_BAUD;
-                    Thread.Sleep(40);
-                }
-                catch { /* best-effort */ }
-            }
-            base.closePort();
-        }
-
-        // =====================================================================
-        // Serial helpers
-        // =====================================================================
-        void DrainInput(int quietMs = 150, int hardLimitMs = 750)
-        {
-            var total = Stopwatch.StartNew();
-            var quiet = Stopwatch.StartNew();
-            while (total.ElapsedMilliseconds < hardLimitMs)
-            {
-                int waiting = serial.BytesToRead;
-                if (waiting > 0)
-                {
-                    var buf = new byte[waiting];
-                    serial.Read(buf, 0, buf.Length);
-                    quiet.Restart();
-                    continue;
-                }
-                if (quiet.ElapsedMilliseconds >= quietMs) break;
-                Thread.Sleep(10);
-            }
-        }
-
-        static string FormatHexSnippet(byte[] data, int maxBytes = 16)
-        {
-            if (data == null || data.Length == 0)
-                return "<none>";
-
-            int shown = data.Length > maxBytes ? maxBytes : data.Length;
-            var sb = new StringBuilder(shown * 3 + 8);
-            for (int i = 0; i < shown; i++)
-            {
-                if (i > 0) sb.Append(' ');
-                sb.Append(data[i].ToString("X2"));
-            }
-            if (data.Length > shown)
-                sb.Append(" ...");
-            return sb.ToString();
-        }
-
-        byte[] ReadExact(int count, int timeoutMs)
-        {
-            var buf = new byte[count];
-            int got = 0;
-            var sw  = Stopwatch.StartNew();
-            while (got < count && sw.ElapsedMilliseconds < timeoutMs && !isCancelled)
-            {
-                try
-                {
-                    int r = serial.Read(buf, got, count - got);
-                    if (r > 0) { got += r; continue; }
-                }
-                catch (TimeoutException) { }
-                Thread.Sleep(5);
-            }
-            if (got == count) return buf;
-            if (got == 0)     return null;
-            var partial = new byte[got];
-            Buffer.BlockCopy(buf, 0, partial, 0, got);
-            return partial;
-        }
-
-        // Best-effort software jump into download mode using:
-        // LF, "upgrade", LF NUL. Devices without console support may ignore
-        // it; the normal 0x55/"OK" sync still follows.
-        void TryEnterDownloadModeByUpgradeCommand()
-        {
-            if (serial == null || !serial.IsOpen)
-                return;
-
-            try
-            {
-                if (serial.BaudRate != XR_ROM_BAUD)
-                {
-                    serial.BaudRate = XR_ROM_BAUD;
-                    Thread.Sleep(30);
-                }
-
-                addLogLine("Attempting XR809 software upgrade command before sync...");
-                DrainInput(quietMs: 40, hardLimitMs: 120);
-
-                serial.Write(new byte[] { 0x0A }, 0, 1);
-                byte[] cmd = Encoding.ASCII.GetBytes("upgrade");
-                serial.Write(cmd, 0, cmd.Length);
-                serial.Write(new byte[] { 0x0A, 0x00 }, 0, 2);
-                serial.BaseStream.Flush();
-                Thread.Sleep(XR_UPGRADE_SETTLE_MS);
-
-                // The running app may echo console text or briefly emit its own
-                // response before dropping into BROM. Clear any such bytes so the
-                // normal 0x55/"OK" sync starts from a clean state.
-                DrainInput(quietMs: 40, hardLimitMs: 200);
-            }
-            catch (Exception ex)
-            {
-                addLogLine($"Upgrade pre-sync command did not complete: {ex.Message}");
-            }
-        }
-
-        // =====================================================================
-        // Sync
-        //
-        // Sends 0x55, expects "OK" (2 bytes) back, up to XR_SYNC_RETRY_COUNT attempts.
-        // =====================================================================
-        bool Sync()
-        {
-            for (int attempt = 1; attempt <= XR_SYNC_RETRY_COUNT; attempt++)
-            {
-                if (isCancelled) return false;
-                DrainInput();
-                serial.Write(new byte[] { 0x55 }, 0, 1);
-                serial.BaseStream.Flush();
-
-                var reply = new byte[2];
-                int got   = 0;
-                var sw    = Stopwatch.StartNew();
-                while (sw.ElapsedMilliseconds < 1500 && got < 2 && !isCancelled)
-                {
-                    try { got += serial.Read(reply, got, 2 - got); }
-                    catch (TimeoutException) { }
-                }
-
-                if (got == 2 && reply[0] == 'O' && reply[1] == 'K')
-                {
-                    // Only mention the attempt number if it wasn't the first try,
-                    // so a normal first-attempt fail (port not yet settled) stays silent.
-                    if (attempt > 1)
-                        addLogLine($"Sync OK (attempt {attempt})");
-                    else
-                        addLogLine("Sync OK");
-                    DrainInput();
-                    return true;
-                }
-            }
-            addErrorLine($"Sync failed after {XR_SYNC_RETRY_COUNT} attempts.");
-            return false;
-        }
-
-        // =====================================================================
-        // Low-level packet exchange
-        // =====================================================================
-        BROMResponse ReadBromResponse(int headerTimeoutMs, int payloadTimeoutMs)
-        {
-            byte[] header = ReadExact(12, headerTimeoutMs);
-            if (header == null || header.Length < 12)
-                throw new IOException(
-                    $"BROM header incomplete ({header?.Length ?? 0}/12 bytes): {FormatHexSnippet(header)}.");
-            if (header[0] != 'B' || header[1] != 'R' || header[2] != 'O' || header[3] != 'M')
-                throw new IOException(
-                    $"BROM magic mismatch: {header[0]:X2} {header[1]:X2} {header[2]:X2} {header[3]:X2}");
-
-            var resp = new BROMResponse
-            {
-                Flags         = header[4],
-                BromVersion   = header[5],
-                Checksum      = (ushort)((header[6] << 8) | header[7]),
-                PayloadLength = (header[8] << 24) | (header[9] << 16) | (header[10] << 8) | header[11],
-                Payload       = new byte[0],
-            };
-
-            if (resp.PayloadLength < 0)
-                throw new IOException($"Negative BROM payload length {resp.PayloadLength}.");
-            if (resp.PayloadLength > XR_MAX_BROM_PAYLOAD)
-                throw new IOException(
-                    $"BROM payload length {resp.PayloadLength} exceeds XR809 transport ceiling {XR_MAX_BROM_PAYLOAD}.");
-
-            // Response checksum is not validated; the BROM itself never checks it either.
-            if (resp.PayloadLength > 0)
-            {
-                byte[] payload = ReadExact(resp.PayloadLength, payloadTimeoutMs);
-                if (payload == null || payload.Length < resp.PayloadLength)
-                    throw new IOException(
-                        $"BROM payload incomplete ({payload?.Length ?? 0}/{resp.PayloadLength} bytes).");
-                resp.Payload = payload;
-            }
-            return resp;
-        }
-
-        BROMResponse ReadFixedHeaderResponse(int headerTimeoutMs)
-        {
-            byte[] header = ReadExact(12, headerTimeoutMs);
-            if (header == null || header.Length < 12)
-                throw new IOException(
-                    $"BROM header incomplete ({header?.Length ?? 0}/12 bytes): {FormatHexSnippet(header)}.");
-            if (header[0] != 'B' || header[1] != 'R' || header[2] != 'O' || header[3] != 'M')
-                throw new IOException(
-                    $"BROM magic mismatch: {header[0]:X2} {header[1]:X2} {header[2]:X2} {header[3]:X2}");
-
-            return new BROMResponse
-            {
-                Flags         = header[4],
-                BromVersion   = header[5],
-                Checksum      = (ushort)((header[6] << 8) | header[7]),
-                PayloadLength = (header[8] << 24) | (header[9] << 16) | (header[10] << 8) | header[11],
-                Payload       = new byte[0],
-            };
-        }
-
-        BROMResponse ExecuteRawPacket(
-            byte[] pkt,
-            int    headerTimeoutMs  = 1500,
-            int    payloadTimeoutMs = 3000,
-            bool   sleepBeforeRead  = false)
-        {
-            serial.Write(pkt, 0, pkt.Length);
-            serial.BaseStream.Flush();
-
-            // The BROM requires a 50 ms delay between sending a read command and
-            // reading back the response header.
-            if (sleepBeforeRead)
-                Thread.Sleep(50);
-
-            return ReadBromResponse(headerTimeoutMs, payloadTimeoutMs);
-        }
-
-        // =====================================================================
-        // GetChipType  (opcode 0x17)
-        //
-        // Sends a 13-byte packet (same format as GetFlashId, opcode 0x17 instead
-        // of 0x18) and reads a single-byte chip type value from the response
-        // payload. Some XR809 ROM/stub revisions return this before the RAM
-        // stub upload and again once the active transport is up.
-        // Failure is silently ignored — not all BROMs respond.
-        // =====================================================================
-        void ReadChipType()
-        {
-            // Packet: BROM magic | flags 0x04 0x00 | checksum 0x60 0x52 | logicalLen 0x00 0x00 0x00 0x01 | opcode 0x17
-            var pkt = new byte[] { 0x42, 0x52, 0x4F, 0x4D, 0x04, 0x00, 0x60, 0x52,
-                                   0x00, 0x00, 0x00, 0x01, 0x17 };
-            try
-            {
-                BROMResponse resp = ExecuteRawPacket(pkt, headerTimeoutMs: 1500,
-                                                          payloadTimeoutMs: 1500);
-                if (!resp.IsError && resp.PayloadLength >= 1)
-                    addLogLine($"Chip type    : {resp.Payload[0]}");
-            }
-            catch { }
         }
 
         // =====================================================================
@@ -743,7 +324,7 @@ namespace BK7231Flasher
         // 6. GetFlashId again so later read/write/erase logic sees the active
         //    stub/BROM capabilities.
         // =====================================================================
-        bool EnsureConnectedAndIdentified()
+        protected override bool EnsureConnectedAndIdentified()
         {
             bromVersion = 0;
             flashID = null;
@@ -1017,7 +598,7 @@ namespace BK7231Flasher
             return resp.Payload;
         }
 
-        byte[] ReadSectors(int sectorIndex, int sectorCount)
+        protected override byte[] ReadSectors(int sectorIndex, int sectorCount)
         {
             Exception lastEx = null;
             for (int attempt = 1; attempt <= XR_READ_RETRY_COUNT && !isCancelled; attempt++)
@@ -1117,7 +698,7 @@ namespace BK7231Flasher
             return true;
         }
 
-        bool WriteSectors(int sectorIndex, byte[] data, int sectorCount)
+        protected override bool WriteSectors(int sectorIndex, byte[] data, int sectorCount)
         {
             Exception lastEx = null;
             for (int attempt = 1; attempt <= XR_WRITE_RETRY_COUNT && !isCancelled; attempt++)
@@ -1157,43 +738,6 @@ namespace BK7231Flasher
             if (lastEx != null)
                 throw new IOException($"Write failed at sector {sectorIndex} after {XR_WRITE_RETRY_COUNT} attempts.", lastEx);
             return false;
-        }
-
-        // =====================================================================
-        // High-level read loop
-        // =====================================================================
-        byte[] InternalRead(int startAddress, int length)
-        {
-            var result       = new byte[length];
-            int done         = 0;
-            int totalSectors = (length + XR_SECTOR_SIZE - 1) / XR_SECTOR_SIZE;
-            int startSector  = startAddress / XR_SECTOR_SIZE;
-            int perRead      = 8;
-
-            logger.setProgress(0, totalSectors);
-            logger.setState("Reading...", Color.Transparent);
-            addLog("Reading: ");
-
-            for (int s = 0; s < totalSectors && !isCancelled; )
-            {
-                int count = Math.Min(perRead, totalSectors - s);
-                addLog($"0x{startAddress + s * XR_SECTOR_SIZE:X6}... ");
-                byte[] chunk = ReadSectors(startSector + s, count);
-                if (chunk == null)
-                    throw new IOException(
-                        $"Read failed at sector {startSector + s} " +
-                        $"(0x{startAddress + s * XR_SECTOR_SIZE:X6}).");
-
-                int take = Math.Min(count * XR_SECTOR_SIZE, length - done);
-                Buffer.BlockCopy(chunk, 0, result, done, take);
-                done += take;
-                s    += count;
-                logger.setProgress(s, totalSectors);
-            }
-
-            addLog(Environment.NewLine);
-            logger.setState("Read complete", Color.DarkGreen);
-            return result;
         }
 
         // =====================================================================
@@ -1251,409 +795,14 @@ namespace BK7231Flasher
             return true;
         }
 
-        // =====================================================================
-        // High-level write loop
-        //
-        // Writes in 16 KB chunks (32 sectors).  The last chunk is zero-padded
-        // to a full sector boundary before being sent.
-        // =====================================================================
-        bool InternalWrite(int startAddress, byte[] data)
+        protected override bool PerformPreWriteErase()
         {
-            int total = (data.Length + XR_WRITE_CHUNK_SIZE - 1) / XR_WRITE_CHUNK_SIZE;
-            int done  = 0;
-
-            logger.setProgress(0, total);
-            logger.setState("Writing...", Color.Transparent);
-            addLog("Writing: ");
-
-            for (int offset = 0; offset < data.Length && !isCancelled; )
-            {
-                int chunkBytes = Math.Min(XR_WRITE_CHUNK_SIZE, data.Length - offset);
-
-                byte[] chunk = MiscUtils.subArray(data, offset, chunkBytes);
-                chunk = MiscUtils.padArray(chunk, XR_SECTOR_SIZE);
-
-                int sectorIndex = (startAddress + offset) / XR_SECTOR_SIZE;
-                int sectorCount = chunk.Length / XR_SECTOR_SIZE;
-
-                addLog($"0x{startAddress + offset:X6}... ");
-                if (!WriteSectors(sectorIndex, chunk, sectorCount))
-                {
-                    addLog(Environment.NewLine);
-                    logger.setState("Write failed", Color.Red);
-                    return false;
-                }
-
-                offset += chunkBytes;
-                logger.setProgress(++done, total);
-            }
-
-            addLog(Environment.NewLine);
-            logger.setState("Write complete", Color.DarkGreen);
-            return !isCancelled;
+            return InternalChipErase();
         }
 
-        // =====================================================================
-        // .img parsing and validation
-        //
-        // Each section has a 64-byte header followed by data.  The header
-        // contains the AWIH magic, header and data checksums, data size, load
-        // address, entry point, and the flash offset of the next section
-        // (0xFFFFFFFF = end of chain).
-        // =====================================================================
-        XRSectionHeader ParseSectionHeader(byte[] img, int offset)
+        protected override bool PerformExplicitErase()
         {
-            return new XRSectionHeader
-            {
-                Magic          = BitConverter.ToUInt32(img, offset + 0x00),
-                Version        = BitConverter.ToUInt32(img, offset + 0x04),
-                HeaderChecksum = BitConverter.ToUInt16(img, offset + 0x08),
-                DataChecksum   = BitConverter.ToUInt16(img, offset + 0x0A),
-                DataSize       = BitConverter.ToUInt32(img, offset + 0x0C),
-                LoadAddr       = BitConverter.ToUInt32(img, offset + 0x10),
-                Entry          = BitConverter.ToUInt32(img, offset + 0x14),
-                BodyLen        = BitConverter.ToUInt32(img, offset + 0x18),
-                Attribute      = BitConverter.ToUInt32(img, offset + 0x1C),
-                NextAddr       = BitConverter.ToUInt32(img, offset + 0x20),
-                SectionId      = BitConverter.ToUInt32(img, offset + 0x24),
-            };
-        }
-
-        // Sum of all 16-bit words across the 64-byte header including the stored
-        // checksum itself must equal 0xFFFF.
-        static bool ValidateHeaderChecksum(byte[] img, int offset)
-        {
-            ushort sum = 0;
-            for (int i = 0; i < 0x40; i += 2)
-                sum += (ushort)(img[offset + i] | (img[offset + i + 1] << 8));
-            return sum == 0xFFFF;
-        }
-
-        // stored_checksum + sum(data words) must equal 0xFFFF.
-        static bool ValidateDataChecksum(byte[] img, int dataOffset, int size, ushort stored)
-        {
-            ushort sum = stored;
-            int i = 0;
-            for (; i + 1 < size; i += 2)
-                sum += (ushort)(img[dataOffset + i] | (img[dataOffset + i + 1] << 8));
-            if ((size & 1) != 0)
-                sum += img[dataOffset + size - 1];
-            return sum == 0xFFFF;
-        }
-
-        int ValidateAndLogImgLayout(byte[] img, bool logLayout = true)
-        {
-            int offset  = 0;
-            int count   = 0;
-            int lastEnd = 0;
-            var seenOffsets = new HashSet<int>();
-
-            while (true)
-            {
-                if (!seenOffsets.Add(offset))
-                    throw new InvalidDataException(
-                        $"Section chain loops back to 0x{offset:X}.");
-                if (offset < 0 || offset + 0x40 > img.Length)
-                    throw new InvalidDataException(
-                        $"Section header at 0x{offset:X} is out of file range.");
-
-                XRSectionHeader hdr = ParseSectionHeader(img, offset);
-
-                if (hdr.Magic != 0x48495741)
-                    throw new InvalidDataException(
-                        $"Bad AWIH magic at 0x{offset:X}: 0x{hdr.Magic:X8}");
-                if (!ValidateHeaderChecksum(img, offset))
-                    throw new InvalidDataException(
-                        $"Header checksum mismatch at 0x{offset:X}.");
-
-                int dataOffset = offset + 0x40;
-                int dataSize   = (int)hdr.DataSize;
-
-                if (dataSize < 0)
-                    throw new InvalidDataException(
-                        $"Negative section size at 0x{offset:X}.");
-                if (dataOffset + dataSize > img.Length)
-                    throw new InvalidDataException(
-                        $"Section data at 0x{offset:X} overruns file.");
-                if (!ValidateDataChecksum(img, dataOffset, dataSize, hdr.DataChecksum))
-                    throw new InvalidDataException(
-                        $"Data checksum mismatch at 0x{offset:X}.");
-
-                if (offset < lastEnd)
-                    throw new InvalidDataException(
-                        $"Section at 0x{offset:X} overlaps earlier image data ending at 0x{lastEnd:X}.");
-
-                lastEnd = dataOffset + dataSize;
-
-                if (logLayout)
-                {
-                    string name = SECTION_NAMES.TryGetValue(hdr.SectionId, out string n)
-                                  ? n : $"id_{hdr.SectionId:X8}";
-                    addLogLine($"  [{count}] {name}");
-                    addLogLine($"      offset=0x{offset:X6}  size=0x{dataSize:X6}  load=0x{hdr.LoadAddr:X8}  entry=0x{hdr.Entry:X8}");
-                }
-
-                if (++count > 100)
-                    throw new InvalidDataException("More than 100 sections – likely corrupt image.");
-
-                if (hdr.NextAddr == 0xFFFFFFFF) break;
-
-                if (hdr.NextAddr < (uint)lastEnd)
-                    throw new InvalidDataException(
-                        $"Next section pointer 0x{hdr.NextAddr:X} falls inside current section ending at 0x{lastEnd:X}.");
-                if (hdr.NextAddr > (uint)(img.Length - 0x40))
-                    throw new InvalidDataException(
-                        $"Next section pointer 0x{hdr.NextAddr:X} is out of file range.");
-                if (hdr.NextAddr <= (uint)offset)
-                    throw new InvalidDataException(
-                        $"Next section pointer 0x{hdr.NextAddr:X} does not advance past 0x{offset:X}.");
-
-                offset = (int)hdr.NextAddr;
-            }
-
-            if (logLayout)
-                addLogLine($"  {count} section(s), effective size 0x{lastEnd:X} ({lastEnd} bytes)");
-            return lastEnd;
-        }
-
-        // =====================================================================
-        // Save read result
-        // =====================================================================
-        public override bool saveReadResult(int startOffset)
-        {
-            if (readResult == null) { addErrorLine("No read result to save."); return false; }
-            try
-            {
-                byte[] dat      = readResult.ToArray();
-                string fileName = MiscUtils.formatDateNowFileName("readResult_" + chipType, backupName, "bin");
-                Directory.CreateDirectory("backups");
-                string fullPath = Path.Combine("backups", fileName);
-                File.WriteAllBytes(fullPath, dat);
-                addSuccess($"Saved {dat.Length} bytes → {fullPath}" + Environment.NewLine);
-                logger.onReadResultQIOSaved(dat, "", fullPath);
-                return true;
-            }
-            catch (Exception ex)
-            {
-                addErrorLine("Save failed: " + ex.Message);
-                return false;
-            }
-        }
-
-        // =====================================================================
-        // BaseFlasher virtual overrides
-        // =====================================================================
-
-        // Needed by the Verify button in FormMain.
-        public override byte[] getReadResult()
-        {
-            return readResult?.ToArray();
-        }
-
-        // Easy Flasher's generic raw doWrite() path relies on Beken-specific
-        // RF/partition assumptions, so it is intentionally disabled for XR809.
-        public override void doWrite(int startSector, byte[] data)
-        {
-            addErrorLine("XR809: raw sector write via doWrite() is not supported on this chip.");
-        }
-
-        public override void doRead(int startSector = 0x000, int sectors = 10, bool fullRead = false)
-        {
-            try
-            {
-                if (!DoGenericSetup())               return;
-                if (!EnsureConnectedAndIdentified()) return;
-
-                int startAddr, length;
-                if (fullRead)
-                {
-                    startAddr = 0;
-                    length    = flashSizeBytes;
-                    addLogLine($"Full read: {flashSizeBytes / 0x100000} MB from 0x000000");
-                }
-                else
-                {
-                    startAddr = PublicSectorIndexToByteAddress(startSector);
-                    length    = PublicSectorCountToByteLength(sectors);
-                    addLogLine($"Partial read: 0x{startAddr:X6} + 0x{length:X6} bytes (framework 4 KB units)");
-                }
-
-                byte[] data = InternalRead(startAddr, length);
-                readResult?.Dispose();
-                readResult = new MemoryStream(data);
-            }
-            catch (OperationCanceledException)
-            {
-                addLogLine("Read cancelled.");
-            }
-            catch (Exception ex)
-            {
-                addErrorLine("Read error: " + ex.Message);
-            }
-            finally
-            {
-                closePort();
-            }
-        }
-
-        public override bool doErase(int startSector = 0x000, int sectors = 10, bool bAll = false)
-        {
-            try
-            {
-                if (!DoGenericSetup())               return false;
-                if (!EnsureConnectedAndIdentified()) return false;
-
-                if (bAll)
-                {
-                    addLogLine($"Full erase requested: {flashSizeBytes / 0x100000} MB");
-                }
-                else
-                {
-                    int requestedStartAddr = PublicSectorIndexToByteAddress(startSector);
-                    int requestedLength    = PublicSectorCountToByteLength(sectors);
-                    addWarningLine("XR809 erase is full-chip only; ignoring requested " +
-                                   $"range 0x{requestedStartAddr:X6} + 0x{requestedLength:X6} bytes.");
-                }
-                bool ok = InternalExplicitChipErase();
-                if (ok) addSuccess("Erase complete!" + Environment.NewLine);
-                return ok;
-            }
-            catch (OperationCanceledException)
-            {
-                addLogLine("Erase cancelled.");
-                return false;
-            }
-            catch (Exception ex)
-            {
-                addErrorLine("Erase error: " + ex.Message);
-                return false;
-            }
-            finally
-            {
-                closePort();
-            }
-        }
-
-        public override void doReadAndWrite(
-            int startSector, int sectors, string sourceFileName, WriteMode rwMode)
-        {
-            if (rwMode == WriteMode.OnlyOBKConfig)
-            {
-                addErrorLine("XR809 does not support standalone OBK config writes.");
-                return;
-            }
-
-            try
-            {
-                if (!DoGenericSetup())               return;
-                if (!EnsureConnectedAndIdentified()) return;
-
-                // ── Backup (ReadAndWrite mode) ────────────────────────────────
-                if (rwMode == WriteMode.ReadAndWrite)
-                {
-                    addLogLine($"Backing up full flash ({flashSizeBytes / 0x100000} MB)...");
-                    byte[] backup = InternalRead(0, flashSizeBytes);
-                    if (isCancelled) return;
-                    readResult?.Dispose();
-                    readResult = new MemoryStream(backup);
-                    if (!saveReadResult(0)) return;
-                }
-
-                if (isCancelled) return;
-
-                // ── Write ─────────────────────────────────────────────────────
-                if (rwMode != WriteMode.OnlyWrite && rwMode != WriteMode.ReadAndWrite) return;
-
-                if (string.IsNullOrEmpty(sourceFileName) || !File.Exists(sourceFileName))
-                {
-                    addErrorLine($"Source file not found: {sourceFileName}");
-                    return;
-                }
-
-                addLogLine($"Loading: {sourceFileName}");
-                byte[] fileData           = File.ReadAllBytes(sourceFileName);
-                bool customRawWrite       = IsCustomRawWriteRequest(startSector, sectors, rwMode);
-                int requestedWriteAddress = PublicSectorIndexToByteAddress(startSector);
-                int writeAddress          = requestedWriteAddress;
-                int effectiveLen          = fileData.Length;
-                string ext                = Path.GetExtension(sourceFileName).ToLowerInvariant();
-                bool treatAsXRImage       = false;
-
-                if (customRawWrite)
-                {
-                    addLogLine($"Custom write: raw bytes only, no erase, destination 0x{writeAddress:X6}.");
-                    if (ext != ".bin")
-                        addWarningLine($"Custom write ignores extension \"{ext}\" and writes raw bytes as selected.");
-                }
-                else if (ext == ".img")
-                {
-                    // Validate now (throws on bad checksum/magic) but defer layout log
-                    // to just before write so it doesn't appear before the erase step.
-                    effectiveLen = ValidateAndLogImgLayout(fileData, logLayout: false);
-                    treatAsXRImage = true;
-                    if (requestedWriteAddress != 0)
-                    {
-                        addWarningLine($".img write ignores requested offset 0x{requestedWriteAddress:X6} " +
-                                       "and starts at 0x000000.");
-                    }
-                    writeAddress = 0;
-                    addLogLine($"Validated .img: 0x{effectiveLen:X} bytes, will write to offset 0x000000");
-                }
-                else if (ext == ".bin")
-                {
-                    if (StartsWithAwihMagic(fileData))
-                    {
-                        addLogLine(".bin starts with AWIH magic, but main .bin writes are treated as raw full-flash data.");
-                    }
-                    addWarningLine($"Writing raw .bin at byte offset 0x{writeAddress:X6}.");
-                }
-                else
-                {
-                    addWarningLine($"Unrecognised extension \"{ext}\" – writing raw bytes.");
-                }
-
-                if ((long)writeAddress + effectiveLen > flashSizeBytes)
-                {
-                    addErrorLine($"Write range 0x{writeAddress:X6} + 0x{effectiveLen:X} bytes exceeds flash size " +
-                                 $"({flashSizeBytes / 0x100000} MB).  Aborting.");
-                    return;
-                }
-
-                byte[] toWrite = new byte[effectiveLen];
-                Buffer.BlockCopy(fileData, 0, toWrite, 0, effectiveLen);
-
-                if (!customRawWrite)
-                {
-                    addLogLine("XR809 write path performs a full-chip erase before programming.");
-                    addLogLine($"Post-erase write destination: 0x{writeAddress:X6}");
-                    if (!InternalChipErase()) return;
-                    if (isCancelled) return;
-
-                    // Show layout now, just before writing, so it appears in context
-                    if (treatAsXRImage)
-                    {
-                        addLogLine("Image layout:");
-                        ValidateAndLogImgLayout(fileData, logLayout: true);
-                    }
-                }
-
-                addLogLine(customRawWrite ? "Writing raw custom data..." : "Writing firmware...");
-                if (!InternalWrite(writeAddress, toWrite)) return;
-
-                addSuccess("XR809 flash write complete!" + Environment.NewLine);
-            }
-            catch (OperationCanceledException)
-            {
-                addLogLine("Operation cancelled.");
-            }
-            catch (Exception ex)
-            {
-                addErrorLine("Fatal error: " + ex.Message);
-            }
-            finally
-            {
-                closePort();
-            }
+            return InternalExplicitChipErase();
         }
     }
 }
